Report on files that cannot be approved by current reviewers
Refactor per-file details in to a class to encapsulate the logic about whether
or not the file is approved, and the comment that should be generated for the
file.

Check the PR's reviewers against the file's owners to determine if the
reviewer set is sufficient to approve the file.

Include any files that can't be approved as a new first block in the
comment output.

Add unit tests that check for this, and check the comment output.